requirements.txt
streamlit
python-docx
langchain
Pillow
sqlite3
